+++ src/pages/schemas/page.schema.ts
@@ -64,6 +64,9 @@ blocksPath.discriminator(
   applicationFormBlockSchema,
 );
 
+pageSchema.index({ 'blocks.media.mediaId': 1 }, { sparse: true });
+pageSchema.index({ 'blocks.backgroundMedia.mediaId': 1 }, { sparse: true });
+pageSchema.index({ 'blocks.items.mediaId': 1 }, { sparse: true });
 pageSchema.index({ slug: 1 }, { unique: true });
 pageSchema.index({ domainId: 1, status: 1, updatedAt: -1 });
 pageSchema.index(